Trim filecoin-ffi from pdp binary via build tags
Introduce the 'pdp' Go build tag so 'make pdp' builds without linking
filecoin-ffi or requiring FFI deps installation.

- Tag sealing/PoRep code paths with !pdp (paths proving, indexing market
  tasks, mk20 handlers, dealdata sealing, webrpc sealing/market RPC)
- Add pdp stubs for paths proving and GPU detection
- Split cuhttp.ServiceDeps into cuhttp/servicedeps (no server graph)
- Split deps Verif/Prover setup into deps_seal.go (!pdp)
- PDP admin API routes omit sector/MK20 webrpc surfaces
- make pdp uses -tags pdp and skips filecoin-ffi BUILD_DEPS
- CGO_ENABLED=0 build verified; go list shows no filecoin-ffi in graph
