+++ deps/deps_pdp.go
@@ -1,0 +1,5 @@
+//go:build pdp
+
+package deps
+
+func setDefaultVerifProver(deps *Deps) {}
