+++ lib/dealdata/logging.go
@@ -1,0 +1,5 @@
+package dealdata
+
+import logging "github.com/ipfs/go-log/v2"
+
+var log = logging.Logger("dealdata")
