+++ lib/paths/local_prove_stub.go
@@ -1,0 +1,28 @@
+//go:build pdp
+
+package paths
+
+import (
+	"context"
+
+	"github.com/ipfs/go-cid"
+	"golang.org/x/xerrors"
+
+	"github.com/filecoin-project/go-state-types/abi"
+
+	"github.com/filecoin-project/curio/lib/storiface"
+)
+
+func (st *Local) GenerateSingleVanillaProof(ctx context.Context, minerID abi.ActorID, si storiface.PostSectorChallenge, ppt abi.RegisteredPoStProof) ([]byte, error) {
+	return nil, xerrors.Errorf("GenerateSingleVanillaProof is not available in the pdp build")
+}
+
+func (st *Local) GeneratePoRepVanillaProof(ctx context.Context, sr storiface.SectorRef, sealed, unsealed cid.Cid, ticket abi.SealRandomness, seed abi.InteractiveSealRandomness) ([]byte, error) {
+	return nil, xerrors.Errorf("GeneratePoRepVanillaProof is not available in the pdp build")
+}
+
+func (st *Local) ReadSnapVanillaProof(ctx context.Context, sr storiface.SectorRef) ([]byte, error) {
+	return nil, xerrors.Errorf("ReadSnapVanillaProof is not available in the pdp build")
+}
+
+var _ Store = &Local{}
