+++ lib/paths/remote_prove.go
@@ -1,3 +1,5 @@
+//go:build !pdp
+
 package paths
 
 import (
