+++ market/mk20/logging.go
@@ -1,0 +1,5 @@
+package mk20
+
+import logging "github.com/ipfs/go-log/v2"
+
+var log = logging.Logger("mk20")
