+++ market/mk20/mk20.go
@@ -1,3 +1,5 @@
+//go:build !pdp
+
 package mk20
 
 import (
