+++ market/mk20/mk20_upload.go
@@ -1,3 +1,5 @@
+//go:build !pdp
+
 package mk20
 
 import (
