+++ pdpnode/routes.go
@@ -7,13 +7,13 @@ import (
 	"github.com/snadrus/must"
 	"golang.org/x/xerrors"
 
-	"github.com/filecoin-project/curio/cuhttp"
+	"github.com/filecoin-project/curio/cuhttp/servicedeps"
 	ipni_provider "github.com/filecoin-project/curio/market/ipni/ipni-provider"
 	"github.com/filecoin-project/curio/pdp"
 )
 
 // MountPDPRoutes attaches PDP HTTP routes using an existing IPNI provider.
-func MountPDPRoutes(ctx context.Context, r chi.Router, d *Deps, sd *cuhttp.ServiceDeps, ipp *ipni_provider.Provider) error {
+func MountPDPRoutes(ctx context.Context, r chi.Router, d *Deps, sd *servicedeps.Deps, ipp *ipni_provider.Provider) error {
 	return pdp.MountRoutes(ctx, r, pdp.MountDeps{
 		DB:         d.DB,
 		LocalStore: d.LocalStore,
@@ -25,7 +25,7 @@ func MountPDPRoutes(ctx context.Context, r chi.Router, d *Deps, sd *cuhttp.Servi
 }
 
 // MountPublicRoutes attaches PDP-only public HTTP routes (IPNI + PDP).
-func MountPublicRoutes(ctx context.Context, r chi.Router, d *Deps, sd *cuhttp.ServiceDeps) error {
+func MountPublicRoutes(ctx context.Context, r chi.Router, d *Deps, sd *servicedeps.Deps) error {
 	ipp, err := ipni_provider.NewProvider(d.CurioDeps())
 	if err != nil {
 		return xerrors.Errorf("ipni provider: %w", err)
