+++ tasks/indexing/task_check_indexes.go
@@ -1,3 +1,5 @@
+//go:build !pdp
+
 package indexing
 
 import (
