+++ web/api/routes_pdp.go
@@ -1,6 +1,5 @@
-//go:build !pdp
+//go:build pdp
 
-// Package api provides the HTTP API for the lotus curio web gui.
 package api
 
 import (
@@ -8,12 +7,10 @@ import (
 
 	"github.com/filecoin-project/curio/deps"
 	"github.com/filecoin-project/curio/web/api/config"
-	"github.com/filecoin-project/curio/web/api/sector"
 	"github.com/filecoin-project/curio/web/api/webrpc"
 )
 
 func Routes(r *mux.Router, deps *deps.Deps, debug bool) {
 	webrpc.Routes(r.PathPrefix("/webrpc").Subrouter(), deps, debug)
 	config.Routes(r.PathPrefix("/config").Subrouter(), deps)
-	sector.Routes(r.PathPrefix("/sector").Subrouter(), deps)
 }
