+++ web/api/webrpc/unsealed_check.go
@@ -1,3 +1,5 @@
+//go:build !pdp
+
 package webrpc
 
 import (
