+++ web/api/webrpc/upgrade.go
@@ -1,3 +1,5 @@
+//go:build !pdp
+
 package webrpc
 
 import (
